combine `withSentry.ts` with `withSentryAPI.ts`
